Introduce a field for overriding the profile levels on a per-OS basis.
Some OSes like COS have L2 checks that other OSes designate as L1. In these cases we'd like to be able to generate different profile levels in the full config when the benchmark is being used for COS.

Also change some of the benchmarks used by COS to L2 using this new field.

PiperOrigin-RevId: 426371189
Change-Id: I7e3114f623365b4be2cda41a3f8a7218c03060c5