Use os.ReadDir instead of ioutil.ReadDir.
os.ReadDir was added in Go 1.16. It's faster and allocates less memory because it avoids calling the lstat syscall on every entry.
In my testing environment, this improves the peak memory usage by ~2.5MB and the execution time by ~35%.

The downside is that the minimum version of Go needed to compile localtoast is now 1.16.

PiperOrigin-RevId: 465542009
Change-Id: Icff22f0a5d101f1f628f7ab932260de69163a460